Add Flow repositories to pom.xml (#7)
* Re-upload the upload component with correct directory setting

this is just a re-done about the initial upload form Mikeal, issue #1
the first upload using a incorrect directory with `com.vaadin.ui` as a 
folder name, which causes error import in ecilpse users. now the correct
directory will `com/vaadin/ui`. :)

* Add Flow repositories to pom.xml

This is done because otherwise the parent-pom, flow-component, can't be
resolved.
- Remove the main method from demo class