Rename test file name (#8)
* Re-upload the upload component with correct directory setting

this is just a re-done about the initial upload form Mikeal, issue #1
the first upload using a incorrect directory with `com.vaadin.ui` as a 
folder name, which causes error import in ecilpse users. now the correct
directory will `com/vaadin/ui`. :)

* Merge branch 'master' of https://github.com/vaadin/vaadin-upload-flow

* Merge branch 'master' of https://github.com/vaadin/vaadin-upload-flow

* Merge branch 'master' of https://github.com/vaadin/vaadin-upload-flow

* Rename UploadIT test file

the name used before can cause a problem during the test process and it
is not the standard IT test file name;

update a version number in pom.